requirements.txt
python-dotenv
openai-whisper
requests
litellm
pydub